[Maintainance] Code refactoring for stratgey
