+++ internal/strategy/consistent_hashing.go
@@ -1,0 +1,96 @@
+package strategy
+
+import (
+	"fmt"
+	"sort"
+
+	"github.com/maniSHarma7575/loadbalancer/internal/loadbalancer"
+)
+
+type ConsistentHashingBS struct {
+	Backends []loadbalancer.Backend
+	Slots    []int
+}
+
+func NewConsistentHashingBS(backends []loadbalancer.Backend) *ConsistentHashingBS {
+	strategy := new(ConsistentHashingBS)
+	strategy.Init(backends)
+	return strategy
+}
+
+func (chbs *ConsistentHashingBS) Init(backends []loadbalancer.Backend) {
+	chbs.Slots = []int{}
+	chbs.Backends = []loadbalancer.Backend{}
+
+	for _, backend := range backends {
+		backendHash := hashFn(backend.Stringify()).Int64()
+
+		if len(chbs.Slots) == 0 {
+			chbs.Slots = append(chbs.Slots, int(backendHash))
+			chbs.Backends = append(chbs.Backends, backend)
+			continue
+		}
+
+		index := sort.Search(len(chbs.Slots), func(i int) bool {
+			return chbs.Slots[i] >= int(backendHash)
+		})
+
+		if index == len(chbs.Slots) {
+			chbs.Slots = append(chbs.Slots, int(backendHash))
+		} else {
+			chbs.Slots = append(chbs.Slots[:index+1], chbs.Slots[index:]...)
+			chbs.Slots[index] = int(backendHash)
+		}
+
+		if index == len(chbs.Backends) {
+			chbs.Backends = append(chbs.Backends, backend)
+		} else {
+			chbs.Backends = append(chbs.Backends[:index+1], chbs.Backends[index:]...)
+			chbs.Backends[index] = backend
+		}
+	}
+}
+
+func (chbs *ConsistentHashingBS) RegisterBackend(backend loadbalancer.Backend) {
+	backendHash := hashFn(backend.Stringify()).Int64()
+
+	index := sort.Search(len(chbs.Slots), func(i int) bool {
+		return chbs.Slots[i] >= int(backendHash)
+	})
+
+	if index == len(chbs.Slots) {
+		chbs.Slots = append(chbs.Slots, int(backendHash))
+	} else {
+		chbs.Slots = append(chbs.Slots[:index+1], chbs.Slots[index:]...)
+		chbs.Slots[index] = int(backendHash)
+	}
+
+	if index == len(chbs.Backends) {
+		chbs.Backends = append(chbs.Backends, backend)
+	} else {
+		chbs.Backends = append(chbs.Backends[:index+1], chbs.Backends[index:]...)
+		chbs.Backends[index] = backend
+	}
+}
+
+func (chbs *ConsistentHashingBS) GetNextBackend(req loadbalancer.IncomingReq) loadbalancer.Backend {
+	hash := hashFn(req.GetReqID()).Int64()
+
+	index := sort.Search(len(chbs.Slots), func(i int) bool {
+		return chbs.Slots[i] >= int(hash)
+	})
+
+	return chbs.Backends[index%len(chbs.Backends)]
+}
+
+func (chbs *ConsistentHashingBS) PrintTopology() {
+	index, i := 0, 0
+	for i = 0; i < 19; i++ {
+		if index < len(chbs.Slots) && chbs.Slots[index] == i {
+			fmt.Printf("		[%2d] %s", i, chbs.Backends[index].Stringify())
+			index++
+		} else {
+			fmt.Printf("		[%2d] -", i)
+		}
+	}
+}
